componentize and cleanup
