+++ package.json
@@ -11,11 +11,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@clalarco/vue3-plotly": "^0.1.5",
     "@tailwindcss/vite": "^4.1.18",
     "@vueuse/core": "^14.2.0",
     "chart.js": "^4.5.1",
-    "plotly.js-basic-dist": "^3.3.1",
     "tailwindcss": "^4.1.18",
     "vue": "^3.5.27",
     "vue-chartjs": "^5.3.3",
